requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.4
django-tinymce==3.5.0
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
mysqlclient==2.1.1
Pillow==9.3.0
django-taggit==3.1.0
psycopg2-binary==2.9.5
dj-database-url==1.2.0 
gunicorn==20.1.0
